chore: Move workers to dir
