+++ apps/event-worker/src/index.ts
@@ -2,9 +2,9 @@ import { logger } from "@ctrlplane/logger";
 
 import { register } from "./instrumentation.js";
 import { redis } from "./redis.js";
-import { createReleaseNewVersionWorker } from "./releases/new-version/index.js";
-import { createReleaseVariableChangeWorker } from "./releases/variable-change/index.js";
 import { workers } from "./workers/index.js";
+import { createReleaseNewVersionWorker } from "./workers/releases/new-version/index.js";
+import { createReleaseVariableChangeWorker } from "./workers/releases/variable-change/index.js";
 
 console.log("Registering instrumentation...");
 await register();
