+++ apps/event-worker/src/workers/releases/mutex.ts
@@ -2,7 +2,7 @@ import type { ReleaseRepository } from "@ctrlplane/rule-engine";
 import type { Mutex as RedisMutex } from "redis-semaphore";
 import { Mutex as RedisSemaphoreMutex } from "redis-semaphore";
 
-import { redis } from "../redis.js";
+import { redis } from "../../redis.js";
 
 export class ReleaseRepositoryMutex {
   static async lock(repo: ReleaseRepository) {
