+++ apps/event-worker/src/workers/resource-scan/aws/vpc.ts
@@ -14,7 +14,7 @@ import { logger } from "@ctrlplane/logger";
 import { ReservedMetadataKey } from "@ctrlplane/validators/conditions";
 
 import type { AwsCredentials } from "./aws.js";
-import { omitNullUndefined } from "../../utils.js";
+import { omitNullUndefined } from "../../../utils.js";
 import { assumeRole, assumeWorkspaceRole } from "./aws.js";
 
 const log = logger.child({ label: "resource-scan/aws/vpc" });
