+++ apps/event-worker/src/workers/resource-scan/index.ts
